[MOD] Moved static datatable extensions to Utility
